+++ config/toml.go
@@ -73,6 +73,9 @@ gui_listen_addr = "{{ .BaseConfig.GUIListenAddress }}"
 # Address to listen for API connections
 api_listen_addr = "{{ .BaseConfig.APIListenAddress }}"
 
+# Enable events for API. Slows down node.
+enable_events = {{ .BaseConfig.EnableEvents }}
+
 # If this node is many blocks behind the tip of the chain, FastSync
 # allows them to catchup quickly by downloading blocks in parallel
 # and verifying their commits