+++ eventsdb/amino.go
@@ -1,0 +1,11 @@
+package eventsdb
+
+import "github.com/tendermint/go-amino"
+
+func RegisterAminoEvents(codec *amino.Codec) {
+	codec.RegisterInterface((*Event)(nil), nil)
+	codec.RegisterConcrete(RewardEvent{},
+		"minter/RewardEvent", nil)
+	codec.RegisterConcrete(SlashEvent{},
+		"minter/SlashEvent", nil)
+}